ApiModel/requirements.txt
Flask
py-near
langchain-community 
python-dotenv # Add this line
